SMGCPA: Fix unit tests and SMGEdgeHasValue equals() and isConsistentWith()
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/smg_witness_for_ldv@24238 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/smg/SMGEdgeHasValueTest.java
@@ -49,7 +49,7 @@ public class SMGEdgeHasValueTest {
     SMGEdgeHasValue hv = new SMGEdgeHasValue(mockType, BigInteger.valueOf(32), obj, val);
 
     Assert.assertEquals(obj, hv.getObject());
-    Assert.assertEquals(32, hv.getOffset());
+    Assert.assertEquals(BigInteger.valueOf(32), hv.getOffset());
     Assert.assertEquals(mockType, hv.getType());
     Assert.assertEquals(32, hv.getSizeInBits(MachineModel.LINUX64));
   }
