+++ src/org/sosy_lab/cpachecker/cpa/smg/objects/sll/SMGSingleLinkedListFinderTest.java
@@ -58,7 +58,7 @@ public class SMGSingleLinkedListFinderTest {
     Assert.assertTrue(candidate instanceof SMGSingleLinkedListCandidateSequence);
     SMGSingleLinkedListCandidateSequence sllCandidate = (SMGSingleLinkedListCandidateSequence)candidate;
     Assert.assertEquals(5, sllCandidate.getLength());
-    Assert.assertEquals(64, sllCandidate.getCandidate().getNfo());
+    Assert.assertEquals(BigInteger.valueOf(64), sllCandidate.getCandidate().getNfo());
     SMGRegion expectedStart = (SMGRegion) smg.getPointer(root.getValue()).getObject();
     Assert.assertSame(expectedStart, sllCandidate.getCandidate().getStartObject());
   }
